namespace and scene rename
