+++ Assets/ZombieHunter/MovementSystem/MovementSystemsContainer.cs
@@ -1,6 +1,5 @@
 ﻿using Leopotam.Ecs;
 using UnityEngine;
-using ZombieHunter.MouseLookSystem;
 using ZombieHunter.MovementSystem.Events;
 using ZombieHunter.InputSystem;
 
@@ -14,8 +13,6 @@ namespace ZombieHunter.MovementSystem
                 .Add(new BlockJumpSystem())
                 .Add(new PlayerInputSystem())
                 .Add(new DevInputSystem())
-                //.Add(new MouseLookSystem.MouseLookSystem())
-                //.Add(new MouseInputSystem())
                 .Add(new MovementSystem())
                 .Add(new GroundCheckSystem())
                 .Add(new PlayerJumpSystem());
